[IMP] website_sale: click and drag for the category filmstrip
Backporting of: `7f0d0486c484b0d5d85aea174548ca0cf5b3eba7` to have the
drag scroll behavior on 16.0.

Original commit message:

In order to display all the category of the filmstrip, It was possible
to use either the small bar by clicking on it and dragging, or to use 
the sliding move on the track pad. Since the bar to be dragged is very
small, it was difficult to use it. Being able to click and drag
everywhere on the filmstrip makes the move easier for the user.

task-3378528

closes odoo/odoo#126429

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>