[FIX] sale_stock: base delivered qty on done qty
In some cases, the delivered quantity of a SOL is incorrect

To reproduce the issue:
1. In Settings, enable "Multi Routes"
2. In Routes, unarchive MTO
3. Create a product P:
   - Storable
   - With a vendor
   - Routes: MTO + Buy
4. Create and confirm a SO for 10 x P
5. On the delivery, set the done quantity to 12
6. Validate the delivery
7. Unlock it and edit the done qty: 10
8. Open the SO

Error: The delivered quantity is still 12

Step 5, because the done quantity is greater than the demand, we
create and process a SM for 2 x P (the done quantity of this SM will
be zero as the initial one keeps a done quantity equal to 12).
Moreover, Since the procure methods are different (MTO vs MTS), the
SMs are not merged.

Step 7, we decrease the done quantity of the initial SM, it impacts
its demand (here actually it does nothing as the demand is already
equal to 10), but we don't change the demand of the newly-created SM.
Here is the issue: to compute the delivered qty, we use the demand
of the done SMs: 10 + 2.

OPW-3504138

closes odoo/odoo#137599

X-original-commit: bf173be70daf0c5ef5af7f467b84c76cfcdb332a
Related: odoo/enterprise#48411
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>