[FIX] mail_plugin: do not allow to create a partner from notification email
Purpose
=======
When an email is opened and is from the default notification mechanism,
i.e: coming from something like "notifications@mail.odoo.com", we don't
want to try to match the email address to an existing contact partner.

Instead, we now display a custom message on the mail addon side
mentioning to search the contact manually.

Task-3557545

closes odoo/odoo#139002

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>