[FIX] website_hr_recruitment: make the job location field not editable
This issue arises when attempting to edit the job location field for a remote
job or when there is no value in the job location, try to edit and save record.

Steps to produce :
- Install `website_hr_recruitment` module.
- Open a job position that has a remote job location or no value in the job
  location.
- Open editor > Edit that field and save the record.
- Error will be generated.

Error: int() argument must be a string, a bytes-like object or a real number,
       not 'NoneType'

This issue occurs because here
https://github.com/odoo/odoo/blob/5e0f5c98330ca1b5d0696c109c9e3a2050a6210c/addons/web_editor/models/ir_qweb.py#L215
it will get None and then when it will try to convert in int than error occurs.

after this commit user will not be able to edit the job location field.

sentry-4588277341

closes odoo/odoo#141246

X-original-commit: 485bc890a7e8ebad7e62ebde9496e53bde3798a8
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>