[FIX] web_editor: link_in_selection style outside editable
Before this commit, the CSS properties defined by the
`o_link_in_selection` class could be applied to a link outside the
editable. In particular, the link preview in the LinkDialog widget would
display the visual indicator of a link in selection, even though it is
not editable nor selected (this happens because the link preview
contains the classes of the original link being edited).

As the styling provided by `.o_link_in_selection` is only meant to be
applied to links inside the editable, this commit makes sure the CSS
rule is specific about that.

task-3580546

closes odoo/odoo#145386

Signed-off-by: Nicolas Bayet (nby) <nby@odoo.com>