[FIX] stock: optimize _search_product_qty
The existing `_search_product_qty` was not scaleable for dbs with many
lots, therefore we optimize it to be more clever.

Also add in the missing operator/value checks.

backport of: ee2fe22db8dcc54d05be89184dd96df6aa9d7148

Note that `_read_group` is only available from saas-16.3 onwards, so
this backport has been adjusted for that. Also pot file does not need to
be updated since the strings already exist in the module.

closes odoo/odoo#144199

X-original-commit: 1336ecf08210b04fb4b365d9e3a2bf8ae434de51
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>