[FIX] point_of_sale: maintain manual quantity for lot-tracked refunds
Before this commit, refunding a lot-tracked product with a quantity
greater than one in the POS  would reset the quantity to one. This
occurred even if the user manually set a higher quantity.

This commit fixes the issue by ensuring the manually set quantity for
lot-tracked products is kept during refunds. This enhancement allows
for accurate quantity retention, improving the POS refund functionality.

opw-3568867

closes odoo/odoo#142678

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>