[FIX] mail: fix non deterministic partner im status test
The "change icon on change partner im_status" test sometimes fails.
This test is using `afterNextRender` that is not robust and does not
provide a lot of information for debugging purpose. This PR adapts
this test to use the `contains` helper instead. This might not fix the
test but will at least provide insight on what is wrong.

fixes runbot-53101

closes odoo/odoo#147084

Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>