[FIX] website_event: no overflow x active for mobile on events page
Steps to reproduce:

Go to the Events Page in any db with a mobile.

You will be be able to scroll vertically as expected, but also
horizontally, this is caused by the negative margin we are using in
`<div class="d-flex mx-n3">` which is causing the content to not fully
fit the screen width and triggering the horizontal scroll.

opw-3429395

closes odoo/odoo#147135

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>