[FIX] sale_margin: update purchase_price on copy
When we copy a sales order we were taking margins from the past which
have a high probability of not being valid anymore.

Now, once we copy a sales order, no past value is copied and the
purchase price is computed as new so we get the current margin values.

TT46721

closes odoo/odoo#148156

X-original-commit: ddf9eed69affd854e8ead02afdbf772b0ba9a1d4
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>