[FIX] mail: redirect with cid when not logged in
A model can specify a preferred company (cid) when redirecting someone from a mail
link.

Before this commit, this cid was taken into account when redirecting only if the
user is logged in.

With this commit, even if the user is not logged in, the redirect link will take
that preferred cid into account.

opw-3613144

closes odoo/odoo#149115

X-original-commit: 06983245a23cbc329308141aec67d71b2d66927b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Adrien Milis (miad) <miad@odoo.com>