[FIX] {hr,sale}_timesheet: enforce consistent no timesheets message
Previous PR: https://github.com/odoo/enterprise/pull/23284 updated the
help message for when there are no timesheets in the view, but forgot to
update the community message to be consistent. This inconsistency
between the community Timesheet app and the enterprise one was both
incorrect + causes an issue with translations since the action is
overriden in enterprise, but the action is stored under a community
module. This made it so the ent. help was never exported to the .pot
file and therefore never translated.

To fix this, we make the messages consistent + re-export the pot file.

Additionally, the `sale_timesheet` action appeared to also be using the
old help message, so we also update this.

closes odoo/odoo#150618

Related: odoo/enterprise#54907
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>